+++ tests/Amadeus/Client/RequestCreator/BaseTest.php
@@ -61,7 +61,7 @@ class BaseTest extends BaseTestCase
         $par = new RequestCreatorParams([
             'originatorOfficeId' => 'BRUXXXXXX',
             'receivedFrom' => 'some RF string',
-            'messagesAndVersions' => ['Fare_DisplayCurrencyIATARates' => '12.1']
+            'messagesAndVersions' => ['Fare_DisplayCurrencyIATARates' => ['version' => '12.1', 'wsdl' => 'aabbccdd']]
         ]);
 
         $rq = new Base($par);
@@ -82,7 +82,7 @@ class BaseTest extends BaseTestCase
         $par = new RequestCreatorParams([
             'originatorOfficeId' => 'BRUXXXXXX',
             'receivedFrom' => 'some RF string',
-            'messagesAndVersions' => ['Fare_DisplayCurrencyIATARates' => '12.1']
+            'messagesAndVersions' => ['Fare_DisplayCurrencyIATARates' => ['version' => '12.1', 'wsdl' => 'aabbccdd']]
         ]);
 
         $rq = new Base($par);
@@ -98,7 +98,7 @@ class BaseTest extends BaseTestCase
         $par = new RequestCreatorParams([
             'originatorOfficeId' => 'BRUXXXXXX',
             'receivedFrom' => 'some RF string',
-            'messagesAndVersions' => ['PNR_Retrieve' => '14.2']
+            'messagesAndVersions' => ['PNR_Retrieve' => ['version' => '14.2', 'wsdl' => 'aabbccdd']]
         ]);
 
         $rq = new Base($par);
@@ -125,7 +125,7 @@ class BaseTest extends BaseTestCase
         $par = new RequestCreatorParams([
             'originatorOfficeId' => 'BRUXXXXXX',
             'receivedFrom' => 'some RF string',
-            'messagesAndVersions' => ['PNR_RetrieveAndDisplay' => '12.1']
+            'messagesAndVersions' => ['PNR_RetrieveAndDisplay' => ['version' => '12.1', 'wsdl' => 'aabbccdd']]
         ]);
 
         $rq = new Base($par);
@@ -148,7 +148,7 @@ class BaseTest extends BaseTestCase
         $par = new RequestCreatorParams([
             'originatorOfficeId' => 'BRUXXXXXX',
             'receivedFrom' => 'some RF string',
-            'messagesAndVersions' => ['Offer_VerifyOffer' => '10.1']
+            'messagesAndVersions' => ['Offer_VerifyOffer' => ['version' => '10.1', 'wsdl' => 'aabbccdd']]
         ]);
 
         $rq = new Base($par);
@@ -172,7 +172,7 @@ class BaseTest extends BaseTestCase
         $par = new RequestCreatorParams([
             'originatorOfficeId' => 'BRUXXXXXX',
             'receivedFrom' => 'some RF string',
-            'messagesAndVersions' => ['Queue_List' => '11.1']
+            'messagesAndVersions' => ['Queue_List' => ['version' => '11.1', 'wsdl' => 'aabbccdd']]
         ]);
 
         $rq = new Base($par);
@@ -217,7 +217,7 @@ class BaseTest extends BaseTestCase
         $par = new RequestCreatorParams([
             'originatorOfficeId' => 'BRUXXXXXX',
             'receivedFrom' => 'some RF string',
-            'messagesAndVersions' => ['Fare_InformativePricingWithoutPNR' => '12.3']
+            'messagesAndVersions' => ['Fare_InformativePricingWithoutPNR' => ['version' => '12.3', 'wsdl' => 'aabbccdd']]
         ]);
 
         $rq = new Base($par);
@@ -237,7 +237,7 @@ class BaseTest extends BaseTestCase
         $par = new RequestCreatorParams([
             'originatorOfficeId' => 'BRUXXXXXX',
             'receivedFrom' => 'some RF string',
-            'messagesAndVersions' => ['Fare_InformativeBestPricingWithoutPNR' => '12.3']
+            'messagesAndVersions' => ['Fare_InformativeBestPricingWithoutPNR' => ['version' => '12.3', 'wsdl' => 'aabbccdd']]
         ]);
 
         $rq = new Base($par);
@@ -261,7 +261,7 @@ class BaseTest extends BaseTestCase
         $par = new RequestCreatorParams([
             'originatorOfficeId' => 'BRUXXXXXX',
             'receivedFrom' => 'some RF string',
-            'messagesAndVersions' => ['Fare_PricePNRWithBookingClass' => '07.3']
+            'messagesAndVersions' => ['Fare_PricePNRWithBookingClass' => ['version' => '07.3', 'wsdl' => 'aabbccdd']]
         ]);
 
         $rq = new Base($par);
@@ -300,7 +300,12 @@ class BaseTest extends BaseTestCase
         $par = new RequestCreatorParams([
             'originatorOfficeId' => 'BRUXXXXXX',
             'receivedFrom' => 'some RF string',
-            'messagesAndVersions' => ['Fare_InformativeBestPricingWithoutPNR' => '14.1']
+            'messagesAndVersions' => [
+                'Fare_InformativeBestPricingWithoutPNR' => [
+                    'version' => '14.1',
+                    'wsdl' => 'dc22e4ee'
+                ]
+            ]
         ]);
 
         $rq = new Base($par);